Refactor copy_mail
Rather than copying on a per mailbox basis, it will fetch all
headers from both the source and destination, and skip mail
that exists on ANY mailbox on the destination.

The previous version only allowed duplicate checking inside the
same mailbox.

With a future commit, this method will also allow the program
to resume significantly faster from a broken pipe.